Fix RPi4-64 builds for traefik
Prior to this change, the wget command used to pull the release of
traefik made use of the BUILD_ARCH environment variable, which on a
Raspberry Pi 4 with a 64 bit system was the string "aarch64". Traefik's
releases for the 64-bit ARM architecture however use "arm64", which
resulted in the release file not being found and the addon container
failing to build. This change fixes the issue by using a TRAEFIK_BUILD_ARCH
variable and setting that appropriately based on BUILD_ARCH.

https://github.com/alex3305/hassio-addons/issues/8